+++ src/models/audiotrack.ts
@@ -39,6 +39,10 @@ export type AudioTrack = {
    */
   status?: string | undefined;
   /**
+   * Title of the track.
+   */
+  title?: string | undefined;
+  /**
    * Name of the language in which the subtitles will be generated.
    *
    * @remarks
@@ -63,6 +67,7 @@ export const AudioTrack$inboundSchema: z.ZodMiniType<AudioTrack, unknown> = z
     id: types.optional(types.string()),
     type: types.optional(AudioTrackType$inboundSchema),
     status: types.optional(types.string()),
+    title: types.optional(types.string()),
     languageName: types.optional(types.string()),
     languageCode: types.optional(types.string()),
   });